requirements.txt
pipecat-ai[azure,groq,silero]
fastapi
uvicorn[standard]
python-dotenv
loguru
groq
scipy
numpy


